# docs/swagger.yaml
definitions:
...
  github_com_mr-isik_gatling-backend_internal_domain.Project:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
...
  internal_api_handler.addMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    type: object
...
  internal_api_handler.createAPIKeyRequest:
    properties:
      name:
        type: string
    type: object
...
  internal_api_handler.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_api_handler.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_api_handler.registerRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI-Powered Load Testing Tool Backend API.
  termsOfService: http://swagger.io/terms/
  title: Gatling-Backend API
  version: "1.0"
paths:
  /auth/api-keys:
    post:
      consumes:
      - application/json
      description: Generates a new API Key for the user (shown only once).
      parameters:
      - description: API Key Name
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.createAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create API Key
      tags:
      - Auth
  /auth/api-keys/{id}:
    delete:
      description: Deletes an existing API Key.
      parameters:
      - description: API Key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete API Key
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with email and password, returns a JWT token pair.
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Login
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Generates a new access token using a refresh token.
      parameters:
      - description: Refresh Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Refresh Token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user and returns a JWT token pair.
      parameters:
      - description: User Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register User
      tags:
      - Auth
  /v1/projects:
    get:
      description: Lists all projects owned by the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Project'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Projects
      tags:
      - Project
    post:
      consumes:
      - application/json
      description: Creates a new project.
      parameters:
      - description: Project Data
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_mr-isik_gatling-backend_internal_domain.Project'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create Project
      tags:
      - Project
  /v1/projects/{id}/members:
    post:
      consumes:
      - application/json
      description: Adds a new member to a project.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Member Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handler.addMemberRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add Project Member
      tags:
      - Project
  /v1/projects/{id}/usage:
    get:
      description: Gets the usage statistics (runs, VUs) for a project.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Project Usage
      tags:
      - Project
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"